Plotting refactor: Moved well positioned colorbar function to _plot_commons.py
