+++ src/astrohack/_utils/_tools.py
@@ -1,7 +1,6 @@
 import json
 
 import numpy as np
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 from prettytable import PrettyTable
 from textwrap import fill
 from astropy.coordinates import EarthLocation, AltAz, HADec, SkyCoord
@@ -117,26 +116,6 @@ def _altaz_to_hadec_astropy(az, el, time, x_ant, y_ant, z_ant):
     ha_dec_coor = azel_coor.transform_to(ha_dec_frame)
     
     return ha_dec_coor.ha, ha_dec_coor.dec
-
-
-def _well_positioned_colorbar(ax, fig, image, label, location='right', size='5%', pad=0.05):
-    """
-    Adds a well positioned colorbar to a plot
-    Args:
-        ax: Axes instance to add the colorbar
-        fig: Figure in which the axes are embedded
-        image: The plt.imshow instance associated to the colorbar
-        label: Colorbar label
-        location: Colorbar location
-        size: Colorbar size
-        pad: Colorbar padding
-
-    Returns: the well positioned colorbar
-
-    """
-    divider = make_axes_locatable(ax)
-    cax = divider.append_axes(location, size=size, pad=pad)
-    return fig.colorbar(image, label=label, cax=cax)
 
 
 def _remove_suffix(input_string, suffix):
